Set -fpermissive when compiling with GCC
The GCC compiler throws an error for the following rule in the
standard: "A name N used in a class S shall refer to the same
declaration in its context and when re-evaluated in the completed
scope of S. No diagnostic is required for a violation of this
rule". As can been seen, no diagnostic is required, so other compilers
are free to only give a warning. To ignore this error on GCC we have
to set the fpermissive flag.